기존 알럿 구조 변경
